[TRA-301] remove home page and replace it with explore page
