+++ src/routes/app.constants.ts
@@ -1,6 +1,5 @@
 export const ROUTES = {
   home: '/',
-  explore: '/explore',
   notFound: '/404',
 
   matrix: '/matrix',
