--- src/routes/home/__tests__/home.component.spec.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Home } from '../home.component';
-import { makeContextRenderer } from '../../../shared/utils/testUtils';
-
-describe('Home: Component', () => {
-  const component = () => <Home />;
-  const render = makeContextRenderer(component);
-
-  it('should render correctly', () => {
-    const { container } = render();
-    expect(container.firstChild).toMatchSnapshot();
-  });
-});
