--- src/routes/home/home.messages.ts
@@ -1,26 +1,0 @@
-/* eslint-disable max-len */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  description: {
-    id: 'home.description',
-    defaultMessage:
-      'The radar will let you explore all of the technologies, methods, and tools used here at Apptension and learn more about them. You will also get to see the technologies we have on our radar (pun intended) and, hopefully, get inspired.',
-  },
-  exploreButton: {
-    id: 'home.exploreButton',
-    defaultMessage: 'Explore',
-  },
-  getInTouchButton: {
-    id: 'home.getInTouchButton',
-    defaultMessage: 'Get in touch',
-  },
-  goToMainPageButton: {
-    id: 'home.goToMainPageButton',
-    defaultMessage: 'Go to apptension.com',
-  },
-  goToAdminPanel: {
-    id: 'newEntry.goToAdminPanel',
-    defaultMessage: 'Go to AdminPanel',
-  },
-});
