--- src/routes/home/index.ts
@@ -1,1 +1,0 @@
-export { Home } from './home.component';
